+++ clikt/src/main/kotlin/com/github/ajalt/clikt/parsers/Parser.kt
@@ -96,12 +96,7 @@ internal object Parser {
             // Finalize eager options
             invocationsByOption.forEach { (o, inv) -> if (o is EagerOption) o.finalize(context, inv) }
 
-            // Finalize option groups
-            invocationsByOptionByGroup.forEach { (group, invocations) ->
-                group?.finalize(context, invocations)
-            }
-
-            // Finalize remaining un-grouped options that occurred on the command line
+            // Finalize un-grouped options that occurred on the command line
             invocationsByOptionByGroup[null]?.forEach { (o, inv) -> o.finalize(context, inv) }
 
             // Finalize un-grouped options not provided on the command line so that they can apply default values etc.
@@ -109,6 +104,11 @@ internal object Parser {
                 if (o !is EagerOption && o !in invocationsByOption && (o as? GroupableOption)?.parameterGroup == null) {
                     o.finalize(context, emptyList())
                 }
+            }
+
+            // Finalize option groups after other options so that the groups can their values
+            invocationsByOptionByGroup.forEach { (group, invocations) ->
+                group?.finalize(context, invocations)
             }
 
             // Finalize groups with no invocations